app/requirements.txt
validators==0.18.2
requests==2.27.1
tldextract==3.2.0
python-whois==0.7.3